requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pathlib2==2.3.7
tqdm==4.66.1
streamlit==1.28.0
